Migrate rest_tornado saltnado tests to pytest
Cherry-picking https://github.com/saltstack/salt/pull/69704/
